requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9
python-dateutil>=2.8
backoff>=2.2.1
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pytest>=7.0
pytest-cov>=4.0
